# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StayGuard API
  version: 1.1.0
  summary: 'Anti-loss guesthouse management API: demo scope plus production v1.1 from the design (docs/15).'
  description: "Contract-first source of truth. Prose companion: docs/04-api-contract.md.\nMoney is a whole number of VND.\
    \ Times are RFC 3339 with offset. The API returns codes and\nstructured values, never display text; the web app translates.\n\
    Extensions: `x-story` (owning story), `x-access` (minimum building access on the target\nbuilding: VIEW or EDIT; OWNER\
    \ means owner role), `x-release` (demo | full).\n    x-access values added in 1.1: PUBLIC, ANY, ANY_STAFF, OWNER_OR_MANAGER,\
    \ EDIT_ANY (EDIT on at least one building).\n"
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
- url: /
  description: Same origin (the Go binary serves both the API and the web app)
security:
- bearerAuth: []
...
paths:
...
  /v1/services:
    get:
      tags:
      - Stays
      operationId: listServices
      summary: Extras catalogue (drinks, food, towels) with stock
      x-story: SG-205
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/rooms/{roomId}/stays:
    post:
      tags:
      - Stays
      operationId: createStay
      summary: Check a guest into a vacant room
      description: The server records check-in time; clients cannot supply it.
      x-story: SG-203
      x-access: EDIT
      parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStayRequest'
      responses:
        '201':
          description: Stay created; room is now OCCUPIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stay'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/stays/{stayId}:
    get:
      tags:
      - Stays
      operationId: getStay
      summary: Stay with a live running total
      x-story: SG-203
      x-access: VIEW
      parameters:
      - $ref: '#/components/parameters/StayId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stay'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/stays/{stayId}/extras:
    post:
      tags:
      - Stays
      operationId: addStayExtras
      summary: Add extras to an active stay
      x-story: SG-205
      x-access: EDIT
      parameters:
      - $ref: '#/components/parameters/StayId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddExtrasRequest'
      responses:
        '200':
          description: Updated stay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stay'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/stays/{stayId}/checkout:
    post:
      tags:
      - Stays
      operationId: checkoutStay
      summary: End the stay and produce the final invoice
      description: The server records check-out time. Repeating the request returns the same invoice.
      x-story: SG-205
      x-access: EDIT
      parameters:
      - $ref: '#/components/parameters/StayId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Invoice created; room is now TO_CLEAN once paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/invoices/{invoiceId}/payments:
    post:
      tags:
      - Payments
      operationId: createPayment
      summary: Start a cash or bank-transfer payment for an invoice
      description: 'TRANSFER returns a QR payload for the exact amount, always for the tenant''s own account.

        CASH is recorded on the caller''s open shift.

        '
      x-story: SG-301
      x-access: EDIT
      parameters:
      - $ref: '#/components/parameters/InvoiceId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/payments/{paymentId}:
    get:
      tags:
      - Payments
      operationId: getPayment
      summary: Payment status (polling fallback for the event stream)
      x-story: SG-301
      x-access: VIEW
      parameters:
      - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/demo/payments/{paymentId}/simulate:
    post:
      tags:
      - Payments
      operationId: simulatePaymentReceived
      summary: Simulate the bank reporting the money as received (demo mode only)
      description: 'Returns 404 unless DEMO_MODE is enabled. Creates a payment event exactly like the real

        webhook and runs the same handler.

        '
      x-story: SG-302
      x-access: EDIT
      x-release: demo
      parameters:
      - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Handler result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/owner/overview:
    get:
      tags:
      - Owner
      operationId: getOwnerOverview
      summary: Revenue, cash expected, occupancy, alerts and latest payments for a day
      x-story: SG-403
      x-access: OWNER
      parameters:
      - name: date
        in: query
        required: false
        description: Tenant-local date; defaults to today.
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/auth/sign-out:
    post:
      tags:
      - Auth
      operationId: signOut
      summary: End the current session
      x-story: SG-701
      x-access: ANY
      x-release: full
      responses:
        '204': &id001
          description: Done
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/me/pin:
    put:
      tags:
      - Auth
      operationId: changeMyPin
      summary: Change my PIN (required after a one-time PIN)
      x-story: SG-701
      x-access: ANY
      x-release: full
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePinRequest'
      responses:
        '204': *id001
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /v1/stays:
    get:
      tags:
      - Stays
      operationId: listStays
      summary: Stay history
      x-story: SG-802
      x-access: VIEW
      x-release: full
      description: Receptionists can query any day within the last `frontDeskHistoryDays` (property setting, default 7) in
        buildings with VIEW or EDIT; owners and managers any range. Each item carries guestId indicators only; numbers and
        photos are never returned here.
      parameters:
      - name: date
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Single day; alternative to from and to
      - name: from
        in: query
        required: false
        schema: *id003
      - name: to
        in: query
        required: false
        schema: *id003
      - name: q
        in: query
        required: false
        schema: *id005
        description: Room, guest name or phone
      - name: buildingId
        in: query
        required: false
        schema: *id005
      - name: state
        in: query
        required: false
        schema: *id005
      - name: cursor
        in: query
        required: false
        schema: *id005
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/StayListItem'
                  nextCursor:
                    type:
                    - string
                    - 'null'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /v1/webhooks/bank:
    post:
      tags:
      - Payments
      operationId: receiveBankWebhookLegacy
      summary: 'Deprecated: single bank webhook (replaced by /v1/webhooks/bank/{hookId})'
      description: 'Kept only so existing configurations fail loudly: answers 410 GONE. Use receiveBankWebhook at /v1/webhooks/bank/{hookId}.'
      security:
      - webhookSignature: []
      x-release: full
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankWebhookPayload'
      responses:
        '202':
          description: Accepted (also returned for duplicates)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '410':
          description: GONE
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      deprecated: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC of the raw body
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated UUID. Same key and body returns the first result; same key with a different body returns
        409.
      schema:
        type: string
        format: uuid
...
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    StayId:
      name: stayId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    InvoiceId:
      name: invoiceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    Unauthorized:
      description: Missing or expired credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated but not allowed (for example BUILDING_FORBIDDEN)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found, or not visible to this tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State conflict or idempotency-key reuse
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine code, e.g. ROOM_NOT_VACANT, BUILDING_FORBIDDEN
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            required:
            - field
            - code
            properties:
              field:
                type: string
              code:
                type: string
    Vnd:
      type: integer
      format: int64
      description: Whole Vietnamese dong
...
    LocalizedText:
      type: object
      required:
      - vi
      - en
      properties:
        vi:
          type: string
        en:
          type: string
...
    RentalType:
      type: string
      enum:
      - HOURLY
      - OVERNIGHT
      - DAILY
    StayStatus:
      type: string
      enum:
      - ACTIVE
      - CHECKED_OUT
    PaymentMethod:
      type: string
      enum:
      - CASH
      - TRANSFER
    PaymentStatus:
      type: string
      enum:
      - PENDING
      - PAID
      - EXPIRED
      - MISMATCH
...
    Service:
      type: object
      required:
      - code
      - name
      - price
      - stock
      properties:
        code:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedText'
        price:
          $ref: '#/components/schemas/Vnd'
        stock:
          type: integer
          minimum: 0
        unit:
          type: string
        lowStockAt:
          type: integer
          minimum: 0
        onSale: *id006
        latestUnitCost:
          oneOf:
          - &id011
            $ref: '#/components/schemas/Vnd'
          - type: 'null'
        soldLast7Days: &id010
          type: integer
    CreateStayRequest:
      type: object
      additionalProperties: false
      required:
      - rentalType
      - guestName
      - guestPhone
      - deposit
      properties:
        rentalType:
          $ref: '#/components/schemas/RentalType'
        guestName:
          type: string
          minLength: 1
          maxLength: 120
        guestPhone:
          type: string
          minLength: 6
          maxLength: 20
        idNumber:
          type:
          - string
          - 'null'
          maxLength: 20
          description: Optional national ID; encrypted at rest; never returned to RECEPTIONIST or HOUSEKEEPING
          never returned in full: null
        deposit:
          $ref: '#/components/schemas/Vnd'
        idConsent:
          type: boolean
          description: Required true when idNumber is given
...
    ExtraLine:
      type: object
      required:
      - serviceCode
      - name
      - quantity
      - unitAmount
      - amount
      properties:
        serviceCode:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedText'
        quantity:
          type: integer
          minimum: 1
        unitAmount:
          $ref: '#/components/schemas/Vnd'
        amount:
          $ref: '#/components/schemas/Vnd'
    Quote:
      type: object
      required:
      - asOf
      - stayAmount
      - extrasAmount
      - total
      - depositPaid
      - balanceDue
      - refundDue
      - capped
      - lines
      properties:
        asOf:
          type: string
          format: date-time
        stayAmount:
          $ref: '#/components/schemas/Vnd'
        extrasAmount:
          $ref: '#/components/schemas/Vnd'
        total:
          $ref: '#/components/schemas/Vnd'
        depositPaid:
          $ref: '#/components/schemas/Vnd'
        balanceDue:
          $ref: '#/components/schemas/Vnd'
        refundDue:
          $ref: '#/components/schemas/Vnd'
        capped:
          type: boolean
        lines:
          type: array
          items:
            $ref: '#/components/schemas/BillLine'
    Stay:
      type: object
      required:
      - id
      - roomId
      - roomCode
      - rentalType
      - status
      - checkInAt
      - guestName
      - guestPhone
      - deposit
      - extras
      - quote
      - pricingVersion
      properties:
        id:
          type: string
        roomId:
          type: string
        roomCode:
          type: string
        rentalType:
          $ref: '#/components/schemas/RentalType'
        status:
          $ref: '#/components/schemas/StayStatus'
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type:
          - string
          - 'null'
          format: date-time
        guestName:
          type: string
        guestPhone:
          type: string
        deposit:
          $ref: '#/components/schemas/Vnd'
        extras:
          type: array
          items:
            $ref: '#/components/schemas/ExtraLine'
        quote:
          $ref: '#/components/schemas/Quote'
        pricingVersion:
          type: integer
          description: Rate plan version snapshotted at check-in
        guestId:
          $ref: '#/components/schemas/GuestIdIndicators'
    AddExtrasRequest:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required:
            - serviceCode
            - quantity
            properties:
              serviceCode:
                type: string
              quantity:
                type: integer
                minimum: 1
                maximum: 99
    Invoice:
      type: object
      required:
      - id
      - stayId
      - roomCode
      - status
      - createdAt
      - quote
      properties:
        id:
          type: string
        stayId:
          type: string
        roomCode:
          type: string
        billCode:
          type: string
          description: Short reference such as PH0930A101, used as the transfer note
        status:
          type: string
          enum:
          - OPEN
          - PAID
        createdAt:
          type: string
          format: date-time
        quote:
          $ref: '#/components/schemas/Quote'
    CreatePaymentRequest:
      type: object
      additionalProperties: false
      required:
      - method
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethod'
    PaymentQr:
      type: object
      required:
      - payload
      - accountNoMasked
      - accountName
      - transferNote
      - amount
      properties:
        payload:
          type: string
          description: VietQR (EMVCo) string; the client renders it as a QR image
        accountNoMasked:
          type: string
        accountName:
          type: string
        transferNote:
          type: string
        amount:
          $ref: '#/components/schemas/Vnd'
    Payment:
      type: object
      required:
      - id
      - invoiceId
      - method
      - status
      - amount
      properties:
        id:
          type: string
        invoiceId:
          type: string
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          $ref: '#/components/schemas/Vnd'
        receivedAmount:
          oneOf:
          - $ref: '#/components/schemas/Vnd'
          - type: 'null'
        paidAt:
          type:
          - string
          - 'null'
          format: date-time
        transactionId:
          type:
          - string
          - 'null'
        qr:
          oneOf:
          - $ref: '#/components/schemas/PaymentQr'
          - type: 'null'
    BankWebhookPayload:
      type: object
      description: The SePay webhook body (https://developer.sepay.vn/vi/sepay-webhooks/tich-hop-webhook); unknown fields are ignored.
      properties:
        id:
          type: integer
          description: SePay transaction id, the same on every retry; deduplication key
        gateway:
          type: string
        transactionDate:
          type: string
          description: YYYY-MM-DD HH:mm:ss, Vietnam time
        accountNumber:
          type: string
        subAccount:
          type: string
        code:
          type:
          - string
          - 'null'
          description: Ignored; the bill code is found inside content
        content:
          type: string
          description: Transfer note as typed by the payer
        transferType:
          type: string
          enum:
          - in
          - out
        description:
          type: string
        transferAmount:
          $ref: '#/components/schemas/Vnd'
        accumulated:
          type: integer
        referenceCode:
          type: string
...
    Alert:
      type: object
      required:
      - id
      - kind
      - createdAt
      properties:
        id:
          type: string
        kind:
          $ref: '#/components/schemas/AlertKind'
        createdAt:
          type: string
          format: date-time
        roomCode:
          type:
          - string
          - 'null'
        shiftId:
          type:
          - string
          - 'null'
        stayId:
          type:
          - string
          - 'null'
        actorName:
          type:
          - string
          - 'null'
        amount:
          oneOf:
          - $ref: '#/components/schemas/Vnd'
          - type: 'null'
        details:
          type: object
          additionalProperties:
            type: string
          description: Kind-specific values, for example oldTime and newTime
    PaymentSummary:
      type: object
      required:
      - paymentId
      - roomCode
      - method
      - amount
      - at
      properties:
        paymentId:
          type: string
        roomCode:
          type: string
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          $ref: '#/components/schemas/Vnd'
        at:
          type: string
          format: date-time
    OwnerOverview:
      type: object
      required:
      - date
      - revenueTotal
      - transfersReceived
      - cashExpected
      - byBuilding
      - occupancy
      - alerts
      - latestPayments
      properties:
        date:
          type: string
          format: date
        revenueTotal:
          $ref: '#/components/schemas/Vnd'
        transfersReceived:
          $ref: '#/components/schemas/Vnd'
        cashExpected:
          $ref: '#/components/schemas/Vnd'
        byBuilding:
          type: array
          items:
            type: object
            required:
            - buildingId
            - name
            - revenue
            properties:
              buildingId:
                type: string
              name:
                type: string
              revenue:
                $ref: '#/components/schemas/Vnd'
        occupancy:
          type: object
          required:
          - occupiedRooms
          - totalRooms
          - overdueRooms
          properties:
            occupiedRooms:
              type: integer
            totalRooms:
              type: integer
            overdueRooms:
              type: integer
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        latestPayments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentSummary'
        buildings:
          type: array
          items:
            $ref: '#/components/schemas/BuildingStatus'
        attention:
          type: array
          items:
            $ref: '#/components/schemas/AttentionItem'
...
    ChangePinRequest:
      type: object
      additionalProperties: false
      required:
      - currentPin
      - newPin
      properties:
        currentPin:
          type: string
          pattern: ^[0-9]{6}$
          writeOnly: true
        newPin:
          type: string
          pattern: ^[0-9]{6}$
          writeOnly: true
          description: Not a run (123456) or repeated digit (111111)
...
    StayListItem:
      type: object
      required:
      - id
      - roomCode
      - guestName
      - rentalType
      - checkInAt
      - status
      - guestId
      properties:
        id: *id005
        roomCode: *id005
        guestName: *id005
        rentalType:
          $ref: '#/components/schemas/RentalType'
        checkInAt: *id012
        checkOutAt:
          type:
          - string
          - 'null'
          format: date-time
        total:
          oneOf:
          - *id011
          - type: 'null'
        paymentMethod:
          oneOf:
          - $ref: '#/components/schemas/PaymentMethod'
          - type: 'null'
        state:
          type: string
          enum:
          - IN_STAY
          - PAID
          - UNPAID
          - MISMATCH
          - TIME_EDITED
        status:
          $ref: '#/components/schemas/StayStatus'
        frontDeskName: *id005
        guestId:
          $ref: '#/components/schemas/GuestIdIndicators'
...
    GuestIdIndicators:
      type: object
      required:
      - hasIdNumber
      - hasFrontPhoto
      - hasBackPhoto
      properties:
        hasIdNumber: *id014
        hasFrontPhoto: *id014
        hasBackPhoto: *id014
...